fix: numpad and updates

+++ hooks/index.ts
@@ -4,6 +4,7 @@ import {
   fetchUpdateAsync,
   reloadAsync,
 } from "expo-updates";
+
 import { Platform } from "react-native";
 import { EventRegister } from "react-native-event-listeners";
 import { MMKV } from "react-native-mmkv";
@@ -74,14 +75,17 @@ export const checkForUpdates = async (force?: boolean) => {
     const resolve = await checkForUpdateAsync();
     if (!resolve.isAvailable) return;
     await fetchUpdateAsync();
-    Alert(
-      "Update Available",
-      "An update to the app has been downloaded to your device. Click the Update button to install it, alternatively, it will be installed on the next boot of the app",
-      [
-        { text: "Dismiss" },
-        { text: "Update", onPress: async () => await reloadAsync() },
-      ],
-    );
+    if (force) {
+      await reloadAsync();
+    } else
+      Alert(
+        "Update Available",
+        "An update to the app has been downloaded to your device. Click the Update button to install it, alternatively, it will be installed on the next boot of the app",
+        [
+          { text: "Dismiss" },
+          { text: "Update", onPress: async () => await reloadAsync() },
+        ],
+      );
   } catch { }
 };
